blog/server/requirements.txt
blinker==1.8.2
cffi==1.17.1
click==8.1.8
cryptography==44.0.3
dotenv==0.9.9
flask==3.0.3
flask-sqlalchemy==3.1.1
greenlet==3.1.1
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.6
MarkupSafe==2.1.5
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
sqlalchemy==2.0.40
typing-extensions==4.13.2
werkzeug==3.0.6
zipp==3.20.2
